Fix source-map mapping
The default template for source-map paths is

  'webpack://[namespace]/[resource-path]?[loaders]'

Namespace is output.library, if that's specified, otherwise it
seems to default to the project name in package.json.

vscode's default sourceMapPathMappings take care of that if
namespace is a simple name, but if it contains a '/' we end up
with an extra component in the path, and it fails to find the
source files.

So everything worked until I dropped the library name (which shouldn't
be needed). This drops the [namespace] from the path, so its no
longer an issue, and source-maps work again.