Don't let the printer change Floats to Numbers
In some cases, the estree printer will change a floating point
number with an integer value, to a simple integer representation

One such case is `4e0` which gets translated to `4`. There may be
others. So check the output of the printer, and if the result
looks like an integer, but the original did not, append `f` or `d`
as appropriate.